DocumentQA: Support the document Q&A
This is the first commit to support the document question&answer.
To ask question to the vector data base, the documentation needs to be
transfered to the vector database by the following command:

$ python3 ./apollo.py doc update elon_musk_wiki.txt
You have 1 document(s)
You have roughly 4586 words in your docs
Created a chunk of size 1117, which is longer than the specified 1000
Using OpenAI as the embedding model
Using embedded DuckDB with persistence: data will be stored in: ./chroma

To run the QA bot:
$ python3 ./apollo.py doc run elon_musk_wiki.txt
Using OpenAI as the embedding model
Using embedded DuckDB with persistence: data will be stored in: ./chroma
Ask> who is elon?
 Elon Musk is a business magnate and investor. He is the founder, CEO and chief engineer of SpaceX; angel investor, CEO and product architect of Tesla, Inc.; owner and CEO of Twitter; founder of the Boring Company; co-founder of Neuralink and OpenAI; and president of the philanthropic Musk Foundation. With an estimated net worth of around $192 billion as of March 27, 2023, primarily from his ownership stakes in Tesla and SpaceX, Musk is the second-wealthiest person in the world, according to both the Bloomberg Billionaires Index and Forbes's real-time billionaires list.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>