Rephraser: Reprase a document with various tones
Rephraser is a useful tool to rephrase the writing before sending out
the writing in different scenarios. The current rephraser support two
tones: 1). business 2). casual. By default, the business will be used
for easy daily business article rephrasing. In the following commmands,
I will demonstrate how to use rephraser:

$ cat ~/writing/test
It's a wonderful day. I would like to go out to have a picinic!
$ python3 ./apollo.py rephrase --tone=casual  ~/writing/test
Rephrasing the article /home/gavin/writing/test
You have 1 document(s)
You have roughly 14 words or 18.666666666666664 tokens in your docs

Rephrasing the document in a casual tone:
You have 1 document(s)
You have roughly 128 words or 170.66666666666666 tokens in your docs

"Ah, what a stunningly gorgeous day out! Just looking out my window
already puts me in the mood for a picnic. I think I'll take advantage of
this rare good weather and head out with my friends and family to a
nearby fine spot - it's too nice of a day to waste it indoors! As we set
out, my mind races with excitement, as I know an immensely delightful
outing awaits us. We'll surely have a barrel of fun, so why not make the
most of the day? I think it's the perfect occasion to bring out the
picnic essentials, filling our tummies with luscious goodies and
quenching our thirst with refreshing drinks. All in all, I'm absolutely
certain that this picnic will be a memorable one!"

The rephrased document has been saved in /home/gavin/writing/test.rephrased

$ cat /home/gavin/writing/test.rephrased
Ah, what a stunningly gorgeous day out! Just looking out my window
already puts me in the mood for a picnic. I think I'll take advantage of
this rare good weather and head out with my friends and family to a
nearby fine spot - it's too nice of a day to waste it indoors! As we set
out, my mind races with excitement, as I know an immensely delightful
outing awaits us. We'll surely have a barrel of fun, so why not make the
most of the day? I think it's the perfect occasion to bring out the
picnic essentials, filling our tummies with luscious goodies and
quenching our thirst with refreshing drinks. All in all, I'm absolutely
certain that this picnic will be a memorable one!

"--debug" is an option to enable the verbose mode in the rephrasing
process:

$ python3 ./apollo.py rephrase  --debug ~/writing/test
Rephrasing the article /home/gavin/writing/test
You have 1 document(s)
You have roughly 14 words or 18.666666666666664 tokens in your docs

Rephrasing the document in a business tone:

> Entering new LLMChain chain...
Prompt after formatting:
Rephrase the following article by a polite and professional tone. Please
remember to correct the grammatical errors, use advanced grammatical
clauses, and make it expressive and detailed. Also, please improve the
wording to make it look like highly-educated native speakers' essay. It
is very important that the quality need to be very rigourous and the
legnth needs to be 1.5 times of the origianl number of words. For
example, the origianl legnth is 100 words, and the rephrased version
needs to be at least 150 words. Also, it's wonderful if you can separate
into several paragraphs for better understanding. This is ready for Elon
Musk to review. This should be in the following format:
 Original article:[original content]

 Rephrased result:[rephrased content]

 Example:
 Original article:The candidate has some KVM/Qemu/IOMMU/VIRTIO
experiences and seems ambitious about the developer-type work. The
candidate shows dedication, focus, persistency, and technical depth from
the challenging bug resolution examples. Also, in the APAC region, we
still need more people familiar with the virtualization realm. Based on
the description, I'll rank the candidate between Software Engineer or
Software Engineer II. In the hiring process, we need to confirm that the
candidate would like to work on the interrupt-driven sustaining
engineering type of work (as mentioned in the written interview, the
candidate would like to focus on the virtualization field and thought
Canonical would have free time to do the community work). In the
candidate's experience, only development type of work was demonstrated
and didn't have any customer-facing and pure bug-fixing oriented work -
Some developers would feel this is the support engineer instead of a
software developer.

 Rephrased result:The candidate possesses a wealth of experience in
KVM/Qemu/IOMMU/VIRTIO and demonstrates a keen ambition to engage in
developer-centric work. Their commitment, concentration, persistence,
and technical acumen are evident from the complex bug resolution
examples provided. Furthermore, there is a current need for individuals
with expertise in virtualization within the APAC region. Upon evaluating
the candidate's background and skills, it appears appropriate to
consider them for a position as either a Software Engineer or Software
Engineer II. However, during the hiring process, it is crucial to
ascertain whether the candidate is genuinely interested in engaging with
interrupt-driven sustaining engineering tasks. While the written
interview suggests a desire to concentrate on the virtualization domain
and an assumption that Canonical would provide ample opportunity for
community work, it is essential to verify the candidate's expectations.
It is worth noting that the candidate's experience primarily pertains to
development work, with no explicit mention of customer-facing or
bug-fixing focused tasks. Consequently, some developers might perceive
the role in question to be more akin to that of a support engineer
rather than a software developer. Ensuring alignment between the
candidate's preferences and the position's responsibilities is of utmost
importance in this hiring process.

 Begin!
 Original article: It's a wonderful day. I would like to go out to have a picinic!

 Rephrased result:

> Finished chain.
You have 1 document(s)
You have roughly 20 words or 26.666666666666664 tokens in your docs

"Today is a phenomenal day, and I am eager to capitalize on this delightful opportunity for a leisurely picnic outdoors."

The rephrased document has been saved in /home/gavin/writing/test.rephrased

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>